Create control.py
Initial code: does not include code for calculating angles (law of cosines) or sending and receiving data. Will include y axis after adding that servo motor to the arm.
Will need to update code eventually to match a config file for the arm to make it easily configurable with different arms.